pyproject.toml
[project]
name = "rexer"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.1",
    "loguru>=0.7.3",
    "psutil>=7.2.2",
    "pyinstaller>=6.19.0",
    "pypiwin32>=223",
    "pystray>=0.19.5",
    "pywebview>=6.1",
    "pywin32>=311",
    "sounddevice>=0.5.5",
    "soundfile>=0.13.1",
    "uvicorn>=0.42.0",
    "watchdog>=6.0.0",
    "websockets>=16.0",
]
